requirements.txt
openpyxl==2.5.9
Pillow==5.3.0
pyTelegramBotAPI==3.6.6
pyzbar==0.1.7
lxml==4.2.5
